backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
opencv-python-headless
google-genai
pytesseract
pandas
numpy

